pyproject.toml
[project]
name = "api-retest-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.116.1",
    "loguru>=0.7.3",
    "redis-om>=0.3.5",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
